- Multiplayer Mac <-> Mac - Fixed a bug in NPC Dialogs - Fixed a bug in Text Input - Version number bumped to 0.4

+++ NOLF2/ClientShellDLL/TO2/ScreenHost.cpp
@@ -529,7 +529,7 @@ void CScreenHost::UpdateBandwidth()
 }
 
 
-void CScreenHost::HandleCallback(uint32 dwParam1, uint32 dwParam2)
+void CScreenHost::HandleCallback(uintptr_t dwParam1, uintptr_t dwParam2)
 {
 	switch(dwParam2)
 	{
