+++ NOLF2/ClientShellDLL/TRON/ScreenHost.h
@@ -38,7 +38,7 @@ public:
 	void	ReadyLaunch(LTBOOL bReady);
 
 protected:
-	void	HandleCallback(uint32 dwParam1, uint32 dwParam2);
+	void	HandleCallback(uintptr_t dwParam1, uintptr_t dwParam2);
 
     LTBOOL	LaunchGame();
     LTBOOL	SetService( );
